+++ discovery-provider/plugins/network-monitoring/docker-compose.yaml
@@ -26,40 +26,42 @@ services:
     networks:
       - audius_dev
   
-  grafana:
-    build:
-      context: grafana
-      args:
-        - GRAFANA_VERSION=latest
-        - GF_INSTALL_IMAGE_RENDERER_PLUGIN=true
-    user: 0:0
-    env_file:
-      - grafana/.env
-    links:
-      - prometheus:prometheus
-    ports: 
-      - "3000:3000"
-    restart: always
-    volumes:
-      - ./data/grafana:/var/lib/grafana
-    networks:
-      - audius_dev
+  # grafana:
+  #   build:
+  #     context: grafana
+  #     args:
+  #       - GRAFANA_VERSION=latest
+  #       - GF_INSTALL_IMAGE_RENDERER_PLUGIN=true
+  #   user: 0:0
+  #   env_file:
+  #     - grafana/.env
+  #   links:
+  #     - prometheus:prometheus
+  #   ports: 
+  #     - "3000:3000"
+  #   restart: always
+  #   volumes:
+  #     - ./data/grafana:/var/lib/grafana
+  #   networks:
+  #     - audius_dev
 
-  prometheus:
-    image: prom/prometheus:v2.33.4
-    user: 0:0
-    extra_hosts:
-      - "host.docker.internal:host-gateway"
-    ports:
-      - "9090:9090"
-    volumes:
-      - ./data/prometheus:/prometheus
-      - ./prometheus/config/prometheus.yml:/etc/prometheus/prometheus.yml
-    networks:
-      - audius_dev
+  # prometheus:
+  #   image: prom/prometheus:v2.33.4
+  #   user: 0:0
+  #   extra_hosts:
+  #     - "host.docker.internal:host-gateway"
+  #   ports:
+  #     - "9090:9090"
+  #   volumes:
+  #     - ./data/prometheus:/prometheus
+  #     - ./prometheus/config/prometheus.yml:/etc/prometheus/prometheus.yml
+  #   networks:
+  #     - audius_dev
 
   network-monitoring:
-    build: '.'
+    build: 
+      context: '.'
+      # dockerfile: Dockerfile.test
     depends_on:
       - network-monitoring-db
       - network-monitoring-push-gateway
@@ -67,7 +69,7 @@ services:
       - audius_dev
 
 volumes:
-  postgres_data_network_monitoring_1:
+  postgres_data_network_monitoring_2:
 
 networks:
   audius_dev:
