+++ discovery-provider/plugins/network-monitoring/src/db/index.ts
@@ -1,18 +1,12 @@
 
 import { Sequelize } from 'sequelize'
+
 import { SequelizeStorage, Umzug } from 'umzug'
+import { setupEnv } from '../utils'
 import create_foreign_connection from './migrations/create_foreign_connection'
 import create_tables from './migrations/create_tables'
 
-// FIXME: Pull in env vars more robustly
-const db = {
-  name: process.env['DB_NAME'] || '',
-  host: process.env['DB_HOST'] || '',
-  port: parseInt(process.env['DB_PORT'] || ''),
-  username: process.env['DB_USERNAME'] || '',
-  password: process.env['DB_PASSWORD'] || '',
-  sql_logger: (process.env['SQL_LOGGING'] || '') in ['T', 't', 'True', 'true', '1']
-}
+const { db } = setupEnv()
 
 const dbConnectionPoolMax: number = 100
 
@@ -53,7 +47,7 @@ export const runDBMigrations = async () => {
   }
 }
 
-export async function runMigrations() {
+export const runMigrations = async () => {
   const umzug = new Umzug({
     migrations: [
       create_tables,
@@ -66,7 +60,7 @@ export async function runMigrations() {
   return umzug.up()
 }
 
-export async function clearDatabase() {
+export const clearDatabase = async () => {
   // clear and recreate database schema, which cascades to all tables and rows in tables
   // for use in testing only - will delete all data in the database!!
   await sequelizeConn.query('DROP SCHEMA IF EXISTS public CASCADE')
@@ -76,4 +70,8 @@ export async function clearDatabase() {
 export const connectToDBAndRunMigrations = async () => {
   await verifyDBConnection()
   await runDBMigrations()
+}
+
+export const closeDBConnection = async () => {
+  await sequelizeConn.close()
 }
